Calculator 클래스의 calculate 메서드를 이용하여 App 클래스에서 사칙연산, 결과 값의 저장 및 조회 기능을 추가함

+++ src/step2/Calculator.java
@@ -23,11 +23,11 @@ public class Calculator {
             }
             default -> System.out.println("정확한 연산자를 입력해주세요");
         }
-        resultHistory.add((double) result);
+        resultHistory.add(result);
         return result;
     }
     public List<Double> getResultHistory() {
-        return new ArrayList<>(resultHistory);  // 방어적 복사
+        return new ArrayList<>(resultHistory);
     }
 
     public void removeResult() {
